moved refresh tests into new refresh.unit.test.ts
